Extracted settings to separate file
